@@ -7,7 +7,7 @@ authors = [
 ]
 license = { text = "Apache-2.0" }
 readme = "README.md"
-requires-python = ">=3.8"
+requires-python = ">=3.10"
 keywords = ["asyncio", "ocr", "tesseract"]
 classifiers = [
     "Development Status :: 5 - Production/Stable",
@@ -16,8 +16,6 @@ classifiers = [
     "License :: OSI Approved :: Apache Software License",
     "Operating System :: OS Independent",
     "Programming Language :: Python :: 3 :: Only",
-    "Programming Language :: Python :: 3.8",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
@@ -85,11 +83,14 @@ platform = "linux"
 files = ["aiopytesseract"]
 show_error_context = true
 verbosity = 0
+strict = true
 no_implicit_optional = true
 warn_unused_configs = true
 warn_return_any = true
 warn_unused_ignores = true
 warn_unreachable = true
+warn_redundant_casts = true
+strict_concatenate = true
 check_untyped_defs = true
 extra_checks = true
 strict_equality = true
@@ -100,6 +101,9 @@ disallow_incomplete_defs = true
 disallow_subclassing_any = true
 disallow_untyped_decorators = true
 disallow_any_generics = true
+disallow_any_unimported = true
+disallow_any_decorated = true
+disallow_any_explicit = true
 
 [tool.coverage.run]
 branch = true
@@ -130,7 +134,7 @@ output = "coverage.xml"
 
 [tool.ruff]
 # https://docs.astral.sh/ruff/settings/
-target-version = "py38"
+target-version = "py310"
 line-length = 88
 indent-width = 4
 unsafe-fixes = false
