@@ -1,6 +1,4 @@
-from typing import Dict, Set, Union
-
-from .file_format import FileFormat
+from aiopytesseract.file_format import FileFormat
 
 TESSERACT_CMD: str = "tesseract"
 
@@ -13,7 +11,7 @@ AIOPYTESSERACT_DEFAULT_PSM: int = 3
 AIOPYTESSERACT_DEFAULT_OEM: int = 3
 
 # https://tesseract-ocr.github.io/tessdoc/Data-Files-in-different-versions.html
-TESSERACT_LANGUAGES: Set[str] = {
+TESSERACT_LANGUAGES: set[str] = {
     "afr",
     "amh",
     "ara",
@@ -140,7 +138,7 @@ TESSERACT_LANGUAGES: Set[str] = {
     "yor",
 }
 
-PAGE_SEGMENTATION_MODES: Dict[int, str] = {
+PAGE_SEGMENTATION_MODES: dict[int, str] = {
     0: "Orientation and script detection (OSD) only.",
     1: "Automatic page segmentation with OSD.",
     2: "Automatic page segmentation, but no OSD, or OCR. (not implemented)",
@@ -157,14 +155,14 @@ PAGE_SEGMENTATION_MODES: Dict[int, str] = {
     13: "Raw line. Single text line, bypassing Tesseract-specific hacks.",
 }
 
-OCR_ENGINE_MODES: Dict[int, str] = {
+OCR_ENGINE_MODES: dict[int, str] = {
     0: "Legacy engine only.",
     1: "Neural nets LSTM engine only.",
     2: "Legacy + LSTM engines.",
     3: "Default, based on what is available.",
 }
 
-OUTPUT_FILE_EXTENSIONS: Dict[Union[str, FileFormat], str] = {
+OUTPUT_FILE_EXTENSIONS: dict[str | FileFormat, str] = {
     FileFormat.ALTO: ".xml",
     FileFormat.HOCR: ".hocr",
     FileFormat.PDF: ".pdf",
