@@ -1,6 +1,6 @@
-from .box import Box
-from .data import Data
-from .osd import OSD
-from .parameter import Parameter
+from aiopytesseract.models.box import Box
+from aiopytesseract.models.data import Data
+from aiopytesseract.models.osd import OSD
+from aiopytesseract.models.parameter import Parameter
 
 __all__ = ["OSD", "Box", "Data", "Parameter"]
